Refactored common.ux module, and added some documentation
